implementando pasta public(+ segurança já que todo cod. não será acessado via url) e limpeza de codigo

+++ remover-video.php
@@ -3,7 +3,11 @@
 $dbPath = __DIR__ . '/banco.sqlite';
 $pdo = new PDO("sqlite:$dbPath");
 
-$id = $_GET['id'];
+$id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT);
+if ($id === false || $id === null) {
+    header('Location: /?sucesso=0');
+    exit();
+}
 
 $sql = 'DELETE FROM videos WHERE id = ?';
 $statement = $pdo->prepare($sql);
